// Web/src/app/models/instance-details.model.ts
export class InstanceDetails {
    instanceId: string;
    symbol: string;
    creationTime: string;
    patternId: string;
    timeScale: number;
    budget: number;
    initialBudget: number;
    cleanGains: number;
    partitionSize: number;
    baseAmount: number;
    nPartitions: number;
    partitionLimit: number;
}